Allow error handler to log stream errors
This lets custom error handlers have a chance to handle the error (mainly to log it) which
fixes a bug where stream errors are completely lost if the built-in logger is not turned on.

This change also makes Fastify log stream errors with all of the information of any other error.