Serialize the payload before running onSend hooks
Serialization now happens before the `onSend` hooks, so the hooks receive the serialized payload.

This also adds a feature where an `onSend` hook can replace the payload with `null` to send a response with an empty body without having the `Content-Type` header get set to `0`. This was not possible before because `null` would either get serialized as JSON or be considered an unserialized value and cause an error to be thrown. (It would be preferable to do this by setting the payload to `undefined` in the hook, but `undefined` currently means that the payload was not changed.)

Fixes #672