Create a separate hook runner for onSend hooks
The `onSend` hooks are a special case because they need extra logic for changing the payload. This extra logic is not needed for the other hooks and it makes the other hooks more prone to a bug where the hook runner's internal state will be changed if a hook function returns a promise that resolves to a value that is not `undefined`.

This change adds a second hook runner just for `onSend` hooks that helps simplify the logic for calling the runner (it's no longer necessary to call `.bind(reply)` during a request) and the main hook runner is simplified so that changing its internal state is no longer possible.

This also improves the hooks benchmarks by 450-600 req/sec.