docs: fix header and body generic types
The generic short-hand types follow order `<Query = DefaultQuery, Params = DefaultParams, Headers = DefaultHeaders, Body = DefaultBody>`
but the docs has them listed as `<Query, Params, Body, Headers>`. This caused one of the examples to be incorrect when it assigned `unknown` to the `Headers` type but explained that it was `unknown` on the `Body` type. The body was undefined which mean it would default to `DefaultBody` which is `any`