http2: fix HEAD requests hanging (#2233) (#2239)
Fix HEAD requests hanging in Node version 10 and 13.5.0.
On these requests, the `preValidationCallback` and `preHandlerCallback`
were exiting prematurely because of `reply.raw.finished` being `true`
with the handler not having sent the response yet. Using `writableEnded`
where available and a combined check of the `writable` and `finished`
properties solved this problem.