Move cemremengu to past contributors
Hello everyone,

My sincere apologies for not being able to contribute for a long time and not communicating this earlier. Moving forward, it is best for me to step aside as Fastify welcomes more active contributors. It was a fantastic experience and I learned a lot from this project so thank you so much for giving me the chance and taking your time.

I hope everyone is doing well in these uncertain times. Stay safe an keep up the great work!