Feature: Add support for Typescript typed replies based on the status code.
Currently the type of replies for routes are the amalgamation of all possible
responses.  This leaves much to be desired for Typescript based type checking
of responses.  This PR attempts to remedy this by adding the ability to specify
response type information for every HTTP status code of the route and check that
the data being sent matches the expressed type.

Adds a new sendReplyTyped() method to FastifyReplies which takes a status code
along with a payload.   In Javascript this simply is reply.statusCode().send().

Add support for RepliesTyped in the ReplyPayload interface which allows a separate
Typescript type to be specified per status code.

Adds associated tests and docs