Improve errors handling in ContentTypeParser.run()
Previously, if by error you didn't call the `done` callback, request was left without a reply or error. With this change, I'm allowing when using a sync function to error by throwing an exception, or by returning a false value (maybe we should allow them, as valid bodies?), and also allow to reply the bodies themselves, both directly or inside a Promise as was already being done.

As a side note, I would fully deprecate the `done` callback and left only return a Promise object as the single async mechanism.