Add dynamic contributors' avatars to Readme
Integrate a dynamic solution to fetch contributors' avatars from their GitHub profiles and display them in the Readme. This enhancement celebrates the diverse community contributing to the project, adding a personal touch to the documentation. By showcasing the faces behind the code, we aim to strengthen the sense of community and recognition for our invaluable contributors. Let's make our Readme reflect the vibrant and diverse community that drives the success of our open-source endeavor.

# 5225

Signed-off-by: Aditya Armal <mr.adityaarmal@gmail.com>