Add example mixin system
See #5061 for issue talking about moving away from declaration merging.

This solution is losely based on and an evolution of the system arrived at for my [`umzeption`](https://github.com/voxpelli/umzeption) module, which in turn I extracted into my [personal type helpers module](https://github.com/voxpelli/type-helpers/blob/main/lib/declaration-types.d.ts) and now revamped into a mixin system.

It originally revolved around enabling third-party extendible discriminated unions – similar to basic declaration merging, but utilizing the power of discriminated unions to get around the downsides of declaration merging.

This mixin solution relies a bit less on the discriminated unions aspect and instead adds a new `hasMixin()` helper (type only for now, will essentially be a type alias for `hasPlugin()`) that type guards the properties so that they only appear when the module is explicitly available.

This should be possible to extend with some implicit checks – like if plugin B follows the registration of plugin A, then an implied `hasMixin('A')` should be made.

It should also be possible to add an `assertMixin('A')` to avoid if-statements and simply throw on missing mixins – TypeScript support both types of type narrowing, but for some reason it didn't immediately accept mine now. I personally rarely use the assert variant – I much rather check and throw an error of my own instead, or handles it more gracefully.

There are some tricks in the mixin helpers that I can expand upon, but its getting late here and I want to wrap this up before the end of the day :pray:

Oh – and this approach has 100% backwards compatibility. All existing declaration merging continues to work. It simply adds an improved way of adding types.

And it works 100% with JSDoc, as that's the kind of types I personally do.